chore(ports): move secondary ports
